Moving test helper into root component
