+++ packages/f-registration/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@justeat/f-registration",
   "description": "Fozzie Registration Form Component",
-  "version": "0.29.0",
+  "version": "0.31.0",
   "main": "dist/f-registration.umd.min.js",
   "files": [
     "dist",
@@ -43,7 +43,7 @@
     "extends @justeat/browserslist-config-fozzie"
   ],
   "dependencies": {
-    "@justeat/f-card": "0.6.0",
+    "@justeat/f-card": "^0.6.0",
     "@justeat/f-form-field": "0.5.0",
     "@justeat/f-services": "1.0.0",
     "@justeat/f-vue-icons": "1.0.0-beta.5",
